# docs/swagger.yaml
basePath: /
definitions:
  ant_internal_platform_render.Response:
    properties:
      data: {}
      error:
        type: string
      status:
        type: integer
    type: object
...
  internal_order.SetOrderGroupRequest:
    properties:
      group_id:
        minimum: 1
        type: integer
    required:
    - group_id
    type: object
...
  internal_order.UpdateOrderStatusRequest:
    properties:
      status:
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        type: integer
    required:
    - status
    type: object
  internal_ordergroup.CreateOrderGroupRequest:
    properties:
      label:
        maxLength: 100
        type: string
    type: object
...
  internal_ordergroup.OrderGroup:
    properties:
      app_id:
        type: integer
      created_at:
        type: string
      division_id:
        type: integer
      id:
        type: integer
      label:
        type: string
      orders:
        description: Orders is the member order summary list. Populated on detail
          reads.
        items:
          $ref: '#/definitions/internal_ordergroup.OrderSummary'
        type: array
      orders_count:
        description: OrdersCount is the number of orders in the group.
        type: integer
      status:
        type: integer
      token:
        type: string
      total:
        description: Total is the sum of all member orders' totals. Populated on detail
          reads.
        type: number
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  internal_ordergroup.UpdateOrderGroupRequest:
    properties:
      label:
        maxLength: 100
        type: string
    type: object
  internal_ordergroup.UpdateOrderGroupStatusRequest:
    properties:
      status:
        enum:
        - 1
        - 2
        - 3
        - 4
        type: integer
    required:
    - status
    type: object
...
host: localhost:8082
info:
  contact: {}
  description: This is the ant microservice.
  title: Ant API
  version: "1.0"
paths:
...
  /order-groups:
    get:
      description: Get a summary list of order groups (orders_count only) for the
        authenticated app
      parameters:
      - description: Max items to return (default 50, max 500)
        in: query
        name: limit
        type: integer
      - description: Items to skip (default 0)
        in: query
        name: offset
        type: integer
      - description: Filter by status (1=open, 2=closed, 3=paid, 4=cancelled)
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_ordergroup.OrderGroup'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: List all order groups
      tags:
      - order-groups
    post:
      consumes:
      - application/json
      description: Create an order group (tab) under which multiple orders can be
        clubbed and settled together. A unique token is generated for the group.
      parameters:
      - description: Order group object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_ordergroup.CreateOrderGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_ordergroup.OrderGroup'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Create a new order group
      tags:
      - order-groups
  /order-groups/{id}:
    delete:
      description: Delete an order group. Member orders are detached (kept as standalone
        orders), not deleted.
      parameters:
      - description: Order Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Delete order group by ID
      tags:
      - order-groups
    get:
      description: Get a single order group with its member orders and the combined
        total
      parameters:
      - description: Order Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_ordergroup.OrderGroup'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Get order group by ID
      tags:
      - order-groups
    put:
      consumes:
      - application/json
      description: Update the order group's label. Status is managed via /order-groups/{id}/status.
      parameters:
      - description: Order Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Order group object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_ordergroup.UpdateOrderGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_ordergroup.OrderGroup'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Update order group by ID
      tags:
      - order-groups
  /order-groups/{id}/status:
    patch:
      consumes:
      - application/json
      description: Set the group status (1=open, 2=closed, 3=paid, 4=cancelled)
      parameters:
      - description: Order Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_ordergroup.UpdateOrderGroupStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_ordergroup.OrderGroup'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Set order group status
      tags:
      - order-groups
  /orders:
    get:
      description: Get a summary list of orders (no items, products_count only) for
        the authenticated app
      parameters:
      - description: Max items to return (default 50, max 500)
        in: query
        name: limit
        type: integer
      - description: Items to skip (default 0)
        in: query
        name: offset
        type: integer
      - description: Filter by status (1=pending, 2=confirmed, 3=completed, 4=cancelled,
          5=paid)
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_order.Order'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: List all orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create an order with one or more items. Items snapshot the product
        name, price and chosen attribute values; only active products can be added.
      parameters:
      - description: Order object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_order.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_order.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Create a new order
      tags:
      - orders
  /orders/{id}:
    delete:
      description: Delete an order and all its items
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Delete order by ID
      tags:
      - orders
    get:
      description: Get a single order with its items
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_order.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Get order by ID
      tags:
      - orders
    put:
      consumes:
      - application/json
      description: Atomically update customer details and sync items in one call.
        Items with an id update quantity only (product/attributes immutable); items
        without an id are added; existing items missing from the payload are deleted.
        Status is managed via /orders/{id}/status.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Order object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_order.UpdateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_order.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Update order by ID
      tags:
      - orders
  /orders/{id}/group:
    patch:
      consumes:
      - application/json
      description: Set group_id to attach the order to an order group (tab), or omit/null
        to detach it.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Group assignment
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_order.SetOrderGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_order.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Attach/detach order to a group
      tags:
      - orders
  /orders/{id}/status:
    patch:
      consumes:
      - application/json
      description: Set the order status (1=pending, 2=confirmed, 3=completed, 4=cancelled,
        5=paid)
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_order.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_order.Order'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Set order status
      tags:
      - orders
...
  /public/order-groups:
    post:
      consumes:
      - application/json
      description: Creates a tab (order group) on behalf of a guest-token caller and
        returns its server-minted token, which can be shared so other family members
        place orders under the same tab. Protected by reCAPTCHA and a hidden honeypot
        field.
      parameters:
      - description: Order group object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_ordergroup.CreatePublicOrderGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_ordergroup.OrderGroup'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Create an order group (public)
      tags:
      - Public
  /public/order-groups/{token}:
    get:
      description: Returns the tab identified by the shareable token, including all
        member orders and the combined total. Used by the order-intake page so a family
        member who has the token can view the whole tab.
      parameters:
      - description: Order group token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_ordergroup.OrderGroup'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Get an order group by token (public)
      tags:
      - Public
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...